+++ lib/cores/injection/injection_container.dart
@@ -1,8 +1,8 @@
-import 'package:coffe_menu/features/appClient/data/datasource/appClient_datasource.dart';
-import 'package:coffe_menu/features/appClient/data/repositories/appCient_respository_impl.dart';
-import 'package:coffe_menu/features/appClient/domain/repositories/appClient_respositories.dart';
-import 'package:coffe_menu/features/appClient/domain/useCases/appClientName_useCases.dart';
-import 'package:coffe_menu/features/appClient/domain/useCases/appClient_useCases.dart';
+import 'package:coffe_menu/features/appClient/data/datasource/appclient_datasource.dart';
+import 'package:coffe_menu/features/appClient/data/repositories/appclient_respository_impl.dart';
+import 'package:coffe_menu/features/appClient/domain/repositories/appclient_respositories.dart';
+import 'package:coffe_menu/features/appClient/domain/useCases/appclientname_usecases.dart';
+import 'package:coffe_menu/features/appClient/domain/useCases/appclient_usecases.dart';
 import 'package:coffe_menu/features/auth/data/datasource/auth_datasource.dart';
 import 'package:coffe_menu/features/auth/data/repositories/auth_repository.dart';
 import 'package:coffe_menu/features/auth/domain/repositories/auth_repositories.dart';
