+++ lib/features/appClient/data/repositories/appclient_respository_impl.dart
@@ -1,6 +1,6 @@
-import 'package:coffe_menu/features/appClient/data/datasource/appClient_datasource.dart';
-import 'package:coffe_menu/features/appClient/domain/entities/appClient_entities.dart';
-import 'package:coffe_menu/features/appClient/domain/repositories/appClient_respositories.dart';
+import 'package:coffe_menu/features/appClient/data/datasource/appclient_datasource.dart';
+import 'package:coffe_menu/features/appClient/domain/entities/appclient_entities.dart';
+import 'package:coffe_menu/features/appClient/domain/repositories/appclient_respositories.dart';
 
 class AppcientRespositoryImpl implements AppclientRespositories {
   final AppclientDatasource appclientDatasource;
@@ -13,6 +13,7 @@ class AppcientRespositoryImpl implements AppclientRespositories {
     return models;
   }
 
+  @override
   Future<AppclientEntities> fetchAppClientName({
     required String appTitle,
   }) async {
