+++ lib/features/appClient/domain/entities/appclient_entities.dart
