+++ lib/features/auth/presentation/widgets/loginForm/login_input.dart
@@ -1,5 +1,5 @@
-import 'package:coffe_menu/cores/widgets/inputTextWidget.dart';
-import 'package:coffe_menu/cores/widgets/textWidget.dart';
+import 'package:coffe_menu/cores/widgets/input_text_widget.dart';
+import 'package:coffe_menu/cores/widgets/text_widget.dart';
 import 'package:flutter/material.dart';
 
 /// A stateless widget that represents a login input form with username and password fields.
